Implement "GET" while using SSTables. Use SemaphoreSlim.
I had a bad feeling when using "async" with "lock". Unfortunately, I didn't have the skill to know if that feeling was legit or not.
When executing the client (test) program, I almost always got a SynchronizationError.
After digging into the problem, the culprit was indeed what I thought.
According to this: https://stackoverflow.com/a/52780836, I should use
SemaphoreSlim, Which I did.

As of now the problem doesn't show anymore. Hopefully it's gone.